host_sharing

Module host_sharing

Structs

HostSharingBroker

Used to ensure that host resources are properly reserved before executing a command spec.

HostSharingGuard

A guard for all permits and resources acquired for a HostSharingBroker.acquire request. Keeps the data structures received from semaphores after acquiring. Semaphores are held until this struct is dropped.

RequestedPermits

WeightPercentage

Enums

HostSharingRequirements

Some commands require that we only run one instance of this binary (using an identifier) to check for other instances of the same binary. Some commands required the full host to run, others just dont care. This enum encapsulates all the different scenarios.

HostSharingStrategy

Determines whether a fair or unfair semaphore is used to manage host sharing

WeightClass

This class is intended to represent the resources required by each test. This is then used to map onto resources available on the machine where the tests are run on in order to not saturate the machine and adversely impact testrunning performance and reliability. Testpilot Classic uses the concept of machine cores to manage resources. Testpilot has two types of tests: Normal (1 core) and Heavy (4 cores) where this would run on a machine with typically 24 cores (managed through a semaphore with size 24). Since Testpilot was introduced we have moved to run on Sandcastle machines with 56 cores so we want to move away from the core-analogy and instead use the term “permits” to describe the limited resources available on each machine. More long term we want improve this to also take into account memory usage, cpu usage etc.